Move FireStoreClient class
